start soul stones
# Conflicts:
#	DofusProtocol/Enums/Custom/EffectsEnum.cs
#	Server/Stump.Server.WorldServer/Game/Actors/Fight/MonsterFighter.cs
#	Server/Stump.Server.WorldServer/Game/Actors/Fight/SlaveFighter.cs
#	Server/Stump.Server.WorldServer/Game/Actors/Fight/SummonedClone.cs
#	Server/Stump.Server.WorldServer/Game/Actors/Fight/SummonedMonster.cs
#	Server/Stump.Server.WorldServer/Game/Actors/Fight/SummonedTurret.cs
#	Server/Stump.Server.WorldServer/Game/Actors/Fight/TaxCollectorFighter.cs
#	Server/Stump.Server.WorldServer/Game/Effects/EffectManager.cs
#	Server/Stump.Server.WorldServer/Handlers/Context/RolePlay/ContextRoleplayHandler.cs

+++ Server/Stump.Server.WorldServer/Game/Actors/Fight/SummonedBomb.cs
@@ -136,7 +136,7 @@ namespace Stump.Server.WorldServer.Game.Actors.Fight
 
         public override bool CanMove() => base.CanMove() && MonsterGrade.MovementPoints > 0;
 
-        public override byte Level => (byte)MonsterBombTemplate.Level;
+        public override short Level => (short)MonsterBombTemplate.Level;
 
         public override StatsFields Stats => m_stats;
